feat: add create_dir support for OPFS
