+++ core/src/services/opfs/core.rs
@@ -15,72 +15,24 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 use std::fmt::Debug;
 
 use wasm_bindgen::JsCast;
-use wasm_bindgen::JsValue;
 use wasm_bindgen_futures::JsFuture;
-use web_sys::window;
 use web_sys::File;
-use web_sys::FileSystemDirectoryHandle;
-use web_sys::FileSystemFileHandle;
-use web_sys::FileSystemGetFileOptions;
 use web_sys::FileSystemWritableFileStream;
 
 use crate::Error;
-use crate::ErrorKind;
 use crate::Result;
 
-fn parse_js_error(msg: JsValue) -> Error {
-    Error::new(
-        ErrorKind::Unexpected,
-        msg.as_string().unwrap_or_else(String::new),
-    )
-}
-
-fn dyn_map_err<T: JsCast>(result: Result<JsValue, JsValue>) -> Result<T, Error> {
-    result.and_then(JsCast::dyn_into).map_err(parse_js_error)
-}
-
-async fn get_handle_by_filename(filename: &str) -> Result<FileSystemFileHandle, Error> {
-    let navigator = window().unwrap().navigator();
-    let storage_manager = navigator.storage();
-    let root: FileSystemDirectoryHandle = JsFuture::from(storage_manager.get_directory())
-        .await
-        .and_then(JsCast::dyn_into)
-        .map_err(parse_js_error)?;
-
-    // maybe the option should be exposed?
-    let opt = FileSystemGetFileOptions::new();
-    opt.set_create(true);
-
-    JsFuture::from(root.get_file_handle_with_options(filename, &opt))
-        .await
-        .and_then(JsCast::dyn_into)
-        .map_err(parse_js_error)
-}
+use super::error::*;
+use super::utils::*;
 
 #[derive(Default, Debug)]
 pub struct OpfsCore {}
 
 impl OpfsCore {
+    #[allow(unused)]
     async fn store_file(&self, file_name: &str, content: &[u8]) -> Result<(), Error> {
         let handle = get_handle_by_filename(file_name).await?;
 
@@ -105,6 +57,7 @@ impl OpfsCore {
         Ok(())
     }
 
+    #[allow(unused)]
     async fn read_file(&self, file_name: &str) -> Result<Vec<u8>, Error> {
         let handle = get_handle_by_filename(file_name).await?;
